extract adding id for new user to repository and do id long (in repository used AtomicLong)

+++ src/main/java/by/tms/spring/config/CalcConfig.java
@@ -22,7 +22,7 @@ public class CalcConfig {
     }
 
     @Bean("history")
-    public Map<Integer, List<ExpressionRecord>> history() {
+    public Map<Long, List<ExpressionRecord>> history() {
         return new HashMap<>();
     }
