+++ src/main/java/by/tms/spring/service/UserService.java
@@ -11,7 +11,7 @@ public interface UserService {
 
     User login(String email, String password);
 
-    User logout(int id);
+    User logout(long id);
 
     List<User> findOnlineUsers();
